optimize blast example
